Fix bug in decoder when data multiple of block size
When we had exactly zero bytes of padding, the decoder would still
delete the amount of bytes determined by the first byte of the last
number. To avoid this, the encoder now adds an extra number when padding
is zero. It uses a small float number to distinguish from the integers
used to encode data.